+++ NEWS.md
@@ -1,8 +1,18 @@
-## PrettyCols 1.0.1.9000+ 2023_08_29
+## PrettyCols 1.0.1.9000+ 2024_06_15
+
+### Breaking changes (1.0.1.9002+)
+
+* Argument `name` in `scale_*_*()` functions is now used to define the title of the legend, to make this consistent with other {ggplot2} `scale` functions. If you have previously used e.g. `scale_fill_manual(name = "Bright")` to say you want to use the `"Bright"` palette, this will result in an error. Please use `scale_fill_manual(palette = "Bright")` instead.
+* Argument `legend_title` in `scale_*_*()` functions is deprecated. Please use `name` instead. For now, `legend_title` still works. 
+
+### Non-breaking changes
 
 * Move Python implementation to separate repository
 * Add `scale_*_pretty_div()` diverging scale functions
-* Changes examples to mtcars data
+* Changes examples to `mtcars` data
+* Remove `scale_name` from `discrete_scale()` calls
+* Add {lifecycle} to Imports
+
 
 ## PrettyCols 1.0.1 2023_01_27